Feat :: Emoji Event 로직 구현 #102
Delete, Add Emoji 로직에서 Event 발생시 채팅방에 이벤트 메시지 보내는 로직 추가하였습니다.

+++ src/main/kotlin/com/seugi/api/domain/chat/presentation/websocket/dto/MessageEventDto.kt
@@ -8,5 +8,5 @@ data class MessageEventDto(
     val type: Type? = null,
     val eventList: List<Long> = emptyList(),
     val messageId: String? = null,
-    val emojiId: Long? = null
+    val emojiId: Int? = null
 )
